修改了RpcContent中添加Obvious观察者模式
netty byteBuf 释放内存修改

+++ my-common/my-common-rpc/my-common-rpc-exchange/src/main/java/indi/uhyils/rpc/netty/enums/RpcStatusEnum.java
@@ -8,55 +8,55 @@ public enum RpcStatusEnum {
     /**
      * 成功
      */
-    OK(20),
+    OK((byte) 20),
     /**
      * 客户端超时
      */
-    CLIENT_TIMEOUT(30),
+    CLIENT_TIMEOUT((byte) 30),
     /**
      * 服务端超时
      */
-    SERVER_TIMEOUT(31),
+    SERVER_TIMEOUT((byte) 31),
     /**
      * 请求失败
      */
-    BAD_REQUEST(40),
+    BAD_REQUEST((byte) 40),
     /**
      * 回应失败
      */
-    BAD_RESPONSE(50),
+    BAD_RESPONSE((byte) 50),
     /**
      * 未找到接口
      */
-    SERVICE_NOT_FOUND(60),
+    SERVICE_NOT_FOUND((byte) 60),
     /**
      * 接口错误
      */
-    SERVICE_ERROR(70),
+    SERVICE_ERROR((byte) 70),
     /**
      * 服务端错误
      */
-    SERVER_ERROR(80),
+    SERVER_ERROR((byte) 80),
     /**
      * 客户端错误
      */
-    CLIENT_ERROR(90),
+    CLIENT_ERROR((byte) 90),
     /**
      * 服务器线程池已耗尽错误
      */
-    SERVER_THREADPOOL_EXHAUSTED_ERROR(100);
+    SERVER_THREADPOOL_EXHAUSTED_ERROR((byte) 100);
 
-    private Integer code;
+    private byte code;
 
-    RpcStatusEnum(Integer code) {
+    RpcStatusEnum(byte code) {
         this.code = code;
     }
 
-    public Integer getCode() {
+    public byte getCode() {
         return code;
     }
 
-    public void setCode(Integer code) {
+    public void setCode(byte code) {
         this.code = code;
     }
 }
